+++ components/Form/subscription.js
@@ -1,13 +1,23 @@
-import React from 'react'
+import React from 'react';
 
-function Subscription() {
-    return (
-        <form action='asyncapi-website.netlify.app' name="form 1" method="POST" className='flex sm:flex-col' data-netlify="true">
-            <input type="hidden" name="form-name" value="form 1" />
-            <input type='email' name='email' placeholder='Enter your email' required className='w-[90%] sm:w-full p-4 border border-black' />
-            <button type='submit' className='p-4 bg-black text-white'>Subscribe</button>
-      </form>
-  )
+function Subcription() {
+	return (
+		<div className='mt-[106px] subscription container flex justify-center'>
+			<div className='w-[1024px] min-h-[253px] lg:py-10 lg:w-full flex flex-col items-center mt-[106px]'>
+				<h3 className='text-[32px] text-white lg:text-center'>
+					Subscribe for AACoT’23 updates!
+				</h3>
+				<a href='https://www.asyncapi.com/newsletter' target='_blank' rel="noreferrer">
+					<button
+						type='submit'
+						className='w-[299px] lg:w-full bg-[#875AE2] h-[84px] mt-[40px] text-[24px] text-white rounded-lg lg:mt-6'
+					>
+						Subscribe
+					</button>
+				</a>
+			</div>
+		</div>
+	);
 }
 
-export default Subscription
+export default Subcription;
